45 - chore: Moved query files to sub folders
