+++ src/query/events/fetchEventsAutocomplete.tsx
@@ -1,5 +1,5 @@
-import { serverRoutes } from "../server/server-routes";
-import { IOptionStr } from "../types/IOptionStr";
+import { serverRoutes } from "../../server/server-routes";
+import { IOptionStr } from "../../types/IOptionStr";
 
 export const fetchEventsAutocomplete = async (signal: AbortSignal): Promise<IOptionStr[]> => {
   const baseUrl = import.meta.env.VITE_BASE_URL;
